Support use of "-c" to allow units to be combined with coefficient
This change adds the "-c" option which lets the user combine the units
with the coefficient instead of them being specified as separate
arguments. In implementing this, the existing error handling was
modified so the program terminates with a non-zero status in the event
of an error, and also tries to give a more descriptive error when
parsing values.