Allow units to be combined with coefficient in one argument
This change lets the user combine the units with the coefficient instead
of requiring them to being specified as separate arguments. In
implementing this, the existing error handling was modified so the
program terminates with a non-zero status in the event of an error, and
it also tries to give a more descriptive error when parsing values.